fix: support long ordinal values (#1869)
* fix: support long ordinal values

* docs: clarify minimum ordinal values (#1869)

* test: align long ordinal assertion (#1869)

* test: cover minimum template ordinals (#1869)

* fix: preserve minimum long ordinal rules (#1869)

+++ src/Humanizer/Localisation/Ordinalizers/SuffixOrdinalizer.cs
@@ -19,13 +19,13 @@ class SuffixOrdinalizer(string masculineSuffix, string feminineSuffix, string ne
     /// <summary>
     /// Ordinalizes the number using the masculine suffix by default.
     /// </summary>
-    public override string Convert(int number, string numberString) =>
+    public override string Convert(long number, string numberString) =>
         Convert(number, numberString, GrammaticalGender.Masculine);
 
     /// <summary>
     /// Ordinalizes the number using the suffix for the requested grammatical gender.
     /// </summary>
-    public override string Convert(int number, string numberString, GrammaticalGender gender)
+    public override string Convert(long number, string numberString, GrammaticalGender gender)
     {
         if (zeroAsPlainNumber && number == 0)
         {
